"https://arxiv.org/abs/2002.12499 On Catastrophic Interference in Atari 2600 Games
This paper is interesting. Let me read the Satanophany chapter and then I'll go for it. After that comes the Blazor talk from yesterday.

9:10am. "The Memento agent is far more sample efficient than training the baseline Rainbow agent longer. The Memento agent achieves the same median performance increase in only 5M frames compared to 200M frames for the longer-training Rainbow agent."

9:15am. Yeah, the fact that a single task in RL might have multiple subtasks which cause catastrophic forgetting is what I considered myself. This is not some big insight. It is just that I do not how to resolve the issue.

Reseting from random weights is interesting, but it is hardly the right way to go. Does the brain really do this? I doubt it.

Though I wonder what the Deepmind/OpenAI guys would think about this paper? If they had integrity they should feel shame over throwing so much hardware at the games without resolving the underlying issues. If they want to be seen as the best, then they should do this rather than waiting for somebody else to do it.

9:35am. Though I had been reading that paper, I still dawdled a bit too much. Let me go through the Blazor talk. Then I'll check out that Svelte post. Then comes the MySQL lecture.

9:40am. https://youtu.be/Khn7sDUSEJM?t=934

This is super cool. Yeah, dedicating myself to thoroughly mastering this for the next few years sounds like what I should be doing. Whatever I do, these kinds of UIs is how I should be expressing myself in programming. More than just using this, I should be diving under the hood to understand the secrets of this.

This is real programming.

9:50am. https://youtu.be/Khn7sDUSEJM?t=1605

Interesting that he is talking about gRPC here for efficient client-server communication.

https://youtu.be/Khn7sDUSEJM?t=3309

Blazor for native UIs? That's the future right there. Or part of it. I did not think they would have gone that far, but they are really going all out here.

10:30am. This was great.

10:35am. I am looking at the HN threads and it seems that there is a way to use F# with Blazor. That should be interesting.

https://stackoverflow.blog/2020/02/26/whats-behind-the-hype-about-blazor/

10:40am. I guess I'll save the MySQL lecture for later. Let me read some of this and then I'll watch a talk by Rich Harris.

11:15am. https://svelte.dev/blog/svelte-3-rethinking-reactivity

Let me watch the video accompanying this.

11:20am. https://youtu.be/AdNJ3fydeao?t=532

I actually saw this coming yesterday. In the Vue tutorial, or was it maybe the Svelte tutorial itself, Brad did not need to spread into the array and assign to it. He could have added to an existing array and then just triggered the update by reassigning itself. That would have worked as well.

11:40am. https://youtu.be/AdNJ3fydeao?t=1136

I wonder what talk he is talking about here?

11:55am. https://youtu.be/AdNJ3fydeao?t=1825

Transitions are interesting too. I've been wondering how I would animate moving cards in my GUI. Svelte has a solution.

12:10pm. Uf, that was a lot to digest. Let me have that break. I want to see more Blazor, but I'll leave that for later.

What I need to do is get back to the grind. I am going to go through the MySQL video and then the ASP.NET lecture after that.

I am not sure whether I feel like messing with the legacy stuff now that I know about Blazor, but I should spend some time on it even if I won't use it. I want to at least understand some of it.

1:30pm. Did I forget to commit? Yes, I did. At any rate, the next thing on my TODO list are chores. Then a bit chilling. Then the MySQL lecture. I'll install that studio and start following along. If not, I'll just move to the ASP.NET lecture."