"11:45am. https://mangadex.org/title/28426 Hagure Idol Jigokuhen
More of this is out. Score. This has been on hiatus for too long.

1:50pm. I slacked a bit too much today, but it is fine.

The next on the list are chores, and then the Express video.

The chores will take a while today.

3:05pm. Done with chores. I need some time to cool down first.

3:15pm. Cooled down.

For the rest of the day, my goal will be to go through the Express course. That will mark the beginning of the end of this documentation reading adventure.

3:30pm. Downloading Postman. I saw him use this in the HTTP video.

4:35pm. https://youtu.be/L72fhGm1tfE?list=PLillGF-RfqbYeckUaD1z6nviTp31GLTH8&t=2155

What he is doing here is not good programming practice.

```
app.get("/api/members/:id", (req, res) => {
    const x = req.param("id");
    const l = members.filter((y) => x === y.id.toString())
    if (l.length) res.json(l)
```

It would be better to filter and then check the length. What makes this whole thing a bit complicated is that it is not stated whether ids can be duplicated. If they cannot then...

4:45pm.

```
app.get("/api/members/:id", (req, res) => {
    const x = parseInt(req.param("id"),10);
    const y = members.find(y => x === y.id);
    if (y) res.json(y)
```

Something like this would be good. Also it would be better to parse the ids as ints, but if that fails no doubt an exception will be thrown.

Also `parseInt` requires a base, otherwise the default is hexadecimal. He is probably not expecting that.

This is some of the ways type systems save your ass.

4:45pm. Ah, invalid parses return Nan. Forgot about that.

5pm. https://youtu.be/L72fhGm1tfE?list=PLillGF-RfqbYeckUaD1z6nviTp31GLTH8&t=2521

Damn, will I manage to finish this today? I am learning a lot from this, but I am sensing my focusing slipping.

Well, I should manage it somehow. I can finish two video in one day.

These lectures are pretty good, they are simple enough that I am able to anticipate what he is going to say.

Learning like this is great. Everything on the screen I understand the purpose of and there is a playful atmosphere. This is completely different from the ASP.NET book.

5:25pm. https://youtu.be/L72fhGm1tfE?list=PLillGF-RfqbYeckUaD1z6nviTp31GLTH8&t=2775

Interesting what errors show up. Here he is using `uuid.v4()` which returns a string, but the ids are supposed to be numbers.

No way around it, I have to change the ids to unique ones.

5:45pm. https://youtu.be/L72fhGm1tfE?list=PLillGF-RfqbYeckUaD1z6nviTp31GLTH8&t=3044

Ah, I had enough. Following along is taking too much time and I got the gist of it. Let me just watch to the end here. I am too tired at this point.

https://youtu.be/L72fhGm1tfE?list=PLillGF-RfqbYeckUaD1z6nviTp31GLTH8&t=3085

Ah, once again, just use find and mutate that. The way he is doing would traverse everything needlessly...well, unless he uses `return` to break the loop.

5:55pm. https://youtu.be/L72fhGm1tfE?list=PLillGF-RfqbYeckUaD1z6nviTp31GLTH8&t=3501

Let me stop here. I know there is only 15m left, but that is on HTML templating. I definitely want to leave this for when I am fresh. The fatigue is really hitting me right now. I've really been focusing and playing with this for hours now.

6pm. Phew.

Let's see...once I deal with this...

I'll need to dive a tad deeper and figure out how to make application on the same computer communicate, rather than web servers. I am really quite close to finding what I seek.

The stuff with routing and matching on requests is all that I've seen in ASP.NET. It really confused me greatly what those `unit -> unit` type functions do. As I suspected, anybody with a modicum of experience in webdev would in fact recognize their function.

Everything he has demonstrated to me today will go quite far in understanding other frameworks.

6:10pm. I am just zoning out here right now. Let me close."