"9:15am. I've been up for a bit. Let me continue chilling and then I will start.
I've decided to ditch row polymorphism after all. I just realized that the use case I wanted it for - adding record fields at compile time is not applicable using just the subtyping operators. This is something that should have been obvious, but I had not thought it through until now.

9:20am. Whatever the case, Spiral v0.2 will have to spend some time in the design stage. There are various things I want it to have, but I am not sure how the features will feel until I give it a try.

9:25am. Now, since yesterday was so laborious let me spend some time resting.

https://arxiv.org/abs/2003.00532
High Performance Code Generation in MLIR: An Early Case Study with GEMM

I also want to read this paper.

9:30am. Let me start wasting time.

https://mangadex.org/title/44146/eiyu-oh-bu-wo-kiwameru-tame-tensei-su-soshite-sekai-saikyou-no-minarai-kisi

More of this guilty pleasure is out, and if there is something I like, it is guilt with my pleasure.

10am. Ok, enough. Time for the paper.

After that comes playing with servers. One the first thing I need to figure out how to target the client to an executable. In the sample the server is a `.js` file so I am not sure if simply passing the path to an `.exe` will work. Still, this challenge should not be too difficult. I should be able to surmount it with some study.

10:15am. Oh yes. I just realizing one thing - there is no need for me to get rid of the regular keyword arg patterns. I can compile `a: int b: int` so it matches on `.a_b_, int, int`. This will make things less efficient than having keyword args directly, but it is fine as it will simplify things across the board. I will have to hack the printer so it makes an attempt at printing `.a_b_, int, int` as `a: int b: int`.

10:25am. In the absence of row polymorphism, one thing I could do is add more record syntax. For example `{r with_only a b}` which would return a record with just those fields. I suppose I could use type level magic for the same thing. If I am applying a record to some function then I know what I need to pare down, so I could let the constraint solver tell me the input and the return type and do the operations that I need in bottom-up mode.

10:50am. I skimmed that paper. Mhhh...I really have been slacking this morning. Do I really start programming now, or should I segue straight to breakfast?

11am. https://microsoft.github.io/language-server-protocol/overviews/lsp/overview/

This stuff makes a lot more sense to me now.

11:05am. Let me stop here. Yeah, I know I hadn't exactly excelled this morning - yesterday I was tired and I did so much more, but I need some time to acclimate myself.

When it is time to start, I will focus on one thing only - starting a server and getting some communication going with the client.

I'll start with the server itself and move from there.

Yesterday I did a lot of studying of the source code for the sample server (and a bit of the client.) Today, I will really have to make the first step in mastering them. There is no more looking around for a way to bypass this hurdle. Until it is done, I need to give it my all at scaling the wall directly."