"1:20pm. Done with chores, breakfast and everything else.
Ah, damn it. I wish there was something interesting to hold my attention, but in the absence of that let me finally play with those servers.

My thinking on this is split.

Right now I have the fundamental knowledge that I need for doing interprocess communication (IPC), so I should be able to start right away, but just the thought of having to potentially spend weeks studying the server and client in depth is making me groan in frustration.

Working on language features is what I should be doing, but instead I will be doing what is essentially a subset of webdev and spending my time essentially trying to cut through the bureaucracy imposed on me by VS Code.

1:30pm. Let me gather my thoughts.

Over since I finished the first half of v0.2, for the first two weeks I was looking for shortcuts. Then I made up my mind to study up on webdev.

And now that is done. My plan was to start working on the language server after that.

I am at this point now.

It is true - I understand much more compared to two weeks ago. Understand how servers work.

But I do not quite understand the VS Code related stuff.

Give how huge the LSP related codebases are, it won't be easy at all to build up a mental model of them.

1:40pm. Before I start work on anything, I need to dwell on this just for a bit.

Right now I have a block.

I absolutely want editor support more than anything else for Spiral v0.2. At the end of 2018 I strongly desired this. Every serious language must do this in order to graduate from a plaything to an actual tool.

1:45pm. Yet, I really hate that what I have to do now is trawl through that OO crapbase written in Typescript.

If I hadn't hated OO before, I sure do now.

This interprocess communication thing should be modeled with a bunch of union types being serialized and passed around. This is the only thing I should really have to know, and yet because they did such a shoddy job, now I have to understand this in its entirety.

1:50pm. It is only really at a time like this that I can start to appreciate Haskell's desire to have everything be immutable.

1:55pm. In my own programming I can arrange everything to be like I want, but having to deal with other people's stuff like this..

2pm. I need to make up my mind, right now I am not really dwelling on this thing and am just getting distracted.

Without a doubt, I need to do this.

Though it will be an inefficient way of doing things, and might eat up a month extra than it should, I need to go this extra mile that I resent.

Once I've done this and understand it, only then will I know how to optimize the stuff on the TS side.

2:05pm. Ah, future me. You really owe me one. The things I do for you.

https://www.reddit.com/r/ProgrammingLanguages/comments/f68u50/d_how_to_do_a_language_plugin/fjx7fo9?utm_source=share&utm_medium=web2x

rlp:

```
Hey, it's been a while since you posted this, but in case you are still looking, there's a C# library that does this: https://github.com/OmniSharp/csharp-language-server-protocol

Take a look at the SampleServer, it's pretty simple to get started. You just need to handle various events in the TextDocumentHandler.

The nuget package is https://www.nuget.org/packages/OmniSharp.Extensions.LanguageServer/

I'm not exactly sure how it works from the VS Code side, but I imagine you just start the language server and somehow hook it in to VS Code's language server support.
```

I got this on that thread I posted. I actually had no known about this. It seems this has both a server and a client implementation.

2:10pm. This is worth looking into.

2:15pm. Ok, let me take a break here and I will start work in earnest.

I will make up my mind to cut through the bureaucracy. I will learn what is needed. I won't regret spending my time on this. I will leave the dream of working on the typechecker for the future, and dedicate myself to grinding my way through this.

I admire elegant code, but a real programmer has to hack his way through other people's stuff at some point.

Even if it takes me a month to make the next step, I will do this. There is no way, but to spend my life here."