"5:10pm. Let me gather my thoughts once again.
For the first two weeks of this, I was studying random stuff without success, and in the last two weeks I've acquired the fundamentals that I need.

Now I am looking which way to go.

I've been thinking about it during the break.

I do not know how many lines the VS Code server and client examples have, probably at least 10k, but even so they are the mountain that I need to scale. During the first two weeks I was just too fresh to get anything out of reading the plugin code, but now if I keep at it I should make progress.

I've been thinking about it and I've concluded that if I did what I did during the first two weeks, and combine that with debugging and experimentation I should be able to figure out what that code does.

5:15pm. Even though I made no headway doing that the first two weeks, the way I am now I should be able to attain skill with focused study.

5:20pm. So it is decided - I am going to make the `lsp-sample` and `vscode-languageserver-node` my main goals.

I've never attempted to study something of this scale. 10k lines or more is no joke. You'd imagine there is a better way, but no - this is something I should tackle head one.

To the enterprise guys producing this stuff, 10k lines is like a Hello World program. Restraint when it comes to code size is definitely something that does not enter their considerations.

5:25pm. Very well.

Now before I call it a day, let me just pass in that string and I will debug it for a bit when it inevitably fails.

5:30pm.

```
export declare type ServerOptions = Executable | {
    run: Executable;
    debug: Executable;
} | {
    run: NodeModule;
    debug: NodeModule;
} | NodeModule | (() => Thenable<ChildProcess | StreamInfo | MessageTransports | ChildProcessInfo>);
```

Oh, there is a `Executable` type. Wonderful.

```
export interface Executable {
    command: string;
    args?: string[];
    options?: ExecutableOptions;
}
```

Ok, TS is being nice now. This would be much more of a pain in the ass in raw JS.

5:35pm. Ok, instead of just standing around here, how about I do this?

5:40pm. Success! Though the extension host gets nothing back, I can confirm in the browser that as long as the extension host is running that the server is active.

So that means I got the server to run, just as I said I would.

Let me stop here for the day. Time for lunch."