"8:30am. I am up. I can't believe a third of the month is already over.
I had some time to think about this last night - rather than trying to rewrite everything in F# right away, what I should do is instead rewrite it in TS so that it follows proper programming principles.

Even if it is in TS, if this code was written properly, it would be much easier to grasp.

Before I can even rewrite it in F#, I need to make sure that I understand it. And it would be much more understandable if it was not such a mess in TS.

8:35am. Now let me chill for a bit.

Actually, before I can chill, let me deal with that side gig a bit more.

https://www.hellovinoth.com/ubuntu-14-04-renaming-ethernet-interfaces-from-p1p1%E2%80%B3-to-eth0%E2%80%B3/

Let me rename that network interface. Then...

https://mangadex.org/title/36283/looking-up-to-magical-girls

Finally this shit is out. This is not a guilty pleasure, but a masterpiece.

8:50am. Ugh, the instructions on that page look like such an absolute pain in the ass to follow.

10:05am. https://askubuntu.com/questions/217635/how-to-rename-an-ethernet-interface

I am going to try editing '/etc/udev/rules.d/75-persistent-net-generator.rules' to do what it says here.

I hate Linux so damn much. It is always takes a humiliating song and dance to get anything to work properly on it. I should be starting my day by chilling and then moving to programming, but I am wasting my time with this instead.

10:15am. For some reason it does not seem like the udev rules are doing anything.

10:30am. https://groups.google.com/a/grasehotspot.org/forum/#!topic/grase-hotspot/mz-i-_03YlE

I hadn't known that `sudo update-grub` recreates the `grub.cfg` file. I've been using that command sporadically.

`GRUB_CMDLINE_LINUX="net.ifnames=0 biosdevname=0"`

I keep seeing this as being `GRUB_CMDLINE_LINUX="net.ifnames=1 biosdevname=0"` elsewhere and it is confusing. Let me try the first one.

11:05am. Ok, I understand now. I finally got the hotspot under control.

`GRUB_CMDLINE_LINUX="net.ifnames=0 biosdevname=0"`

Modifying this line it `/etc/default/grub` and then running `update-grub` to recreate the grub config file, and then reseting causes the `/etc/udev/rules.d/70-persistent-net.rules` file to get generated with the appropriate interface cards. Then I can just change the names directly, no need to write all of that shit (that I did like a dozen times) out.

Changing `eth1` to `eth0` for the first card and vice versa for the second gets the thing to work. Well, hopefully it will stand the test when my father gets back.

11:15am. That was such a massive annoyance. My whole morning that I could have spent programming is kaput. The info on SO and elsewhere was all magic incantations without much in the way of explanation of what the various things are supposed to be doing.

Well, if I ever have this problem again, I have the keyword in the journal.

Based on what I wrote last time, I might need to turn off the GRO. I'll cross that bridge later if I need to.

11:20am. Now let me stop here so I can have breakfast. It is time that I read that masterpiece.

When it is time to resume, I will take a look at the rest of the files in the rpcjson project.

I have the basic idea that I need to rewrite this in TS first firmly in mind. One thing I am not sure of yet is whether `messageReader` and `messageWriter` could be implemented in terms of generic streams. And the same question will apply when I do the .NET rewrite.

11:25am. I'll get to the bottom of this today.

I'll crank up the strict typechecking and do this from scratch. By the time I am done, I would not be surprised if I cut the side of the code in half. I will make this trash pile into a gem."