"5:20pm. The way is the market is going is definitely feeling like 2008. At that time though, by the time the crash happened finance stocks were making new lows for 4 quarters straight and bonds were getting crushed, so one had plenty of warning to get out from a technical perspective.
This selloff here happened right in the middle of a bull run in tech stocks.

Even in 1999 it took half a year after the market crashed for the bear market to really start.

At most I can see is the the other sectors were particularly hot, but still there place to be in the stock market is hardly a secret. The stupid thing would have been to own small caps before all of this.

A dumb man sells because stocks are going down, the smart man sells because stocks are not going up.

5:25pm. I have no idea whether I could have anticipated the scale of the selloff from just watching the news. I really thought that 2 days ago was the bottom.

As expected, it is difficult to account for selloffs like these unless you really have a strategy of staying out of the market completely because you are waiting for the big crash to get active. Right now I am definitely in bull mode though. Since the market is in complete demoralization stage, I would not be surprised to see it start rallying today towards the end of the day. If not that then tomorrow.

But it is easy to see this when you are not invested in anything.

5:30pm. Let me resume. Stop lurking the /biz/ sticky.

What is next?

The `server` and the `client`.

5:55pm.

```
arg1 !== void 0
```

Yesterday I said that this tests for `undefined` rather than `null`. Let me try it out in the repl. It might turn out to be the case that it tests for both.

Ah, `void 0` is literally `undefined`. Then I was right.

6:05pm.

```
class TestStream extends Duplex {
 _write(chunk: string, _encoding: string, done: () => void) {
  this.emit('data', chunk);
  done();
 }
 _read(_size: number) { }
}
```

Yeah, he obviously did not know what classes were 3 years ago. Definitely came from a JS background.

6:15pm. Forget it, I do not feel like it anymore.

I did a little skim to see if thing make more sense to me. And they do to an extent. But it is obvious that server just on its own has a massive amount of complexity.

I really am wondering whether I am better just doing it on my own.

Well, I'll have enough to decide it pretty soon. I am nearly done with this now. Having a standard language protocol sounds good at first glance, but if they overload it with features just to cover all the cases then I really am better just rolling my own.

One thing I'll want to do is check out how Purescript does it. Surely it won't let me down. It actually goes as far as advertizing its tooling so that should be something.

6:25pm. It is amazing. Literally everywhere I've looked, I can only find trash.

Let me close here."