"8:55am. I am up.
9:05am. Let me chill for a bit and then I will start.

9:10am. Ok, focus me.

Today my goal will be to finish the book. By the looks of things, it seems it will take me a total of two days before I dive into the language source. That is some kind of record being set.

It feels too fast, but it makes reasonable sense that I take this course so I'll follow along.

9:55am.

```
unsafePartial :: forall a. (Partial => a) -> a
```

I admit, I never ever imagined a type like this. No doubt, PS will push me more towards the pure side of FP language design, but that is fine.

10am.

---

```
showPoint :: Point -> String
showPoint (Point { x, y }) = ...
```

Here, we only specify the names of the properties, and we do not need to specify the names of the values we want to introduce. This is called a record pun.

---

So that is what they call it. It is a good name.

10:10am. Done with chapter 5. I am third of the way into the book. Actually, at this rate I am not sure if I will manage to deal with it all today, but I should be able to do most of it, leaving the rest for tomorrow if needed.

Let me just get on with it.

10:20am. Focus me. Do 3 more chapters of the book and then comes the break.

```
q :: String
q = show $ \n -> n + n // error due to the type var being ambiguous
```

I find it really interesting the way this deals with leftover type variables. This will have a strong benefit when I do this. I admit it can be difficult to think these things through until one just sits and looks at concrete examples - I thought I would have to drag type variables around in constraints, but now I see that it can be done like this.

10:40am. Had to take a break. Let me resume.

11:05am.

```
The inferred type

  forall stream a. Stream stream a => stream -> Maybe stream

has type variables which are not mentioned in the body of the type. Consider adding a type\
 annotation.
```

Here the book talks about multi parameter type classes and the issues with them. Yeah, the above is exactly the problem I ran into while thinking about them myself.

11:20am. Done with chapter 6. Let me do 7 in the next half hour (if possible) and then I will take a break.

11:45am. https://leanpub.com/purescript/read#leanpub-auto-the-eff-monad

I am going through this quickly, but even though I am not familiar with this, given all the experience in Spiral and then Coq/Agda/Lean, this is easy enough for me now. This type level stuff is certainly a piece of cake compared to the Software Foundations proofs.

11:50am. Though I am aiming for a lower level language with Spiral, I really should consider importing as much of Purescript as possible even if it means raising the abstraction bar and having to take extra time implementing things.

11:55am. Ah, I thought that the book had 18 chapters, but it turns out that the total number is 14. I am more further in that I thought.

Let me stop here so I can have breakfast. Things are going well enough."