"8:20am. I am up. Let me chill for a while and then I will start, as is usual.
9:25am. Enough slacking. Let me start.

Let me go through the stuff in the `app` folder.

9:35am. Focus me focus. Let me just go through the `app` folder. The rest can wait for later.

```
        , Opts.command "graph"
            (Opts.info Graph.command
              (Opts.progDesc "Module dependency graph"))
```

Interesting that Purescript has this.

```
        , Opts.command "hierarchy"
            (Opts.info Hierarchy.command
              (Opts.progDesc "Generate a GraphViz directed graph of PureScript type classes"))
```

And this. I do not know what GraphViz is, but the name is evocative enough.

9:55am.

```
compile :: PSCMakeOptions -> IO ()
compile PSCMakeOptions{..} = do
  input <- globWarningOnMisses (unless pscmJSONErrors . warnFileTypeNotFound) pscmInput
  when (null input && not pscmJSONErrors) $ do
    hPutStr stderr $ unlines [ "purs compile: No input files."
                             , "Usage: For basic information, try the `--help' option."
                             ]
    exitFailure
  moduleFiles <- readUTF8FilesT input
  (makeErrors, makeWarnings) <- runMake pscmOpts $ do
    ms <- CST.parseModulesFromFiles id moduleFiles
    let filePathMap = M.fromList $ map (\(fp, pm) -> (P.getModuleName $ CST.resPartial pm, Right fp)) ms
    foreigns <- inferForeignModules filePathMap
    let makeActions = buildMakeActions pscmOutputDir filePathMap foreigns pscmUsePrefix
    P.make makeActions (map snd ms)
  printWarningsAndErrors (P.optionsVerboseErrors pscmOpts) pscmJSONErrors makeWarnings makeErrors
  exitSuccess
```

I am going to have to figure out what every single thing in here does. I am going to have to learn to compile without necessarily relying on Spago.

Well, forget that for a bit here, let me take a look at the rest of the files. After that rather than just trying to read the code, I will have to acclimate myself to doing things by myself. I am going to have to carve out my own subset of the language and study what it does first hand. Everything here is too complex, but bit by bit, I will figure it out.

10:35am. `Ide.hs` is something I am going to have to thoroughly familiarize myself with.

Right now I might just be skimming things, but I am aware that I am not exactly going this the right way. After the initial bit of familiarity is attained, I won't be able to understand more just by looking at the code. I am going to have to play with it.

Whether it be type inference, typeclasses, package management or editor support. I am going to master every single bit of it.

11:15am. I admit, I getting distracted and starting to surf /a/ rather than the thing I am supposed to be doing.

`REPL.hs` is more complex than I thought it would be. In fact, its functionality is on par with the server.

It seems it has been made to run on the browser. That is interesting.

11:25am. I guess it bears a mention, it seems when it comes to IDE support, they decided against a file system watcher in favor of the editor telling the server when to update. I was wondering about this and have no firm views either way. I'll follow Purescript's lead here then.

11:45am. Ok, enough of this.

My focus has really been low. But I think I understand the basic structure of the project.

11:50am. Let me stop here. When it is time to resume what I will do is start looking for ways of paring things down. I need to figure out how to trace through the program - this is something that will be hard to do with Haskell. I'll probably have to insert print statements.

I should make a bunch of simple files and see what the compiler does during their compilation. I will keep increasing the complexity until I have a grasp for all the aspects of the language. At the end of this, I will figure out how the editor support and the REPL works.

Once I have that, I will be at the threshold of being able to start work on Spiral v0.2. I have no idea when that is going to happen. Maybe in May?

There is a lot of work to be done before I can begin, but not an unimaginable amount though."