"9:15am. I am up. I've been getting used to getting up earlier than usual lately, but today it was not the case.
So let me chill for a little.

9:50am. Let me finally start. What is on the table today?

10am. `pathsWithInfo <- traverse (\fp -> (fp,) <$> getInfo fp) inputPaths`

Uf, this kind of heavily monadic code is hard even for me.

...Ok, I get it.

10:05am. No actually I do not.

```
(fp,)
```

What exactly is supposed to go on the right side here?

---

The -XTupleSections flag enables Python-style partially applied tuple constructors. For example, the following program

```
(, True)
```

is considered to be an alternative notation for the more unwieldy alternative

```
\x -> (x, True)
```

---

Ahh, I see. I was wrong about what this does. Clumsy me.

10:10am. I am looking at the documentation for `</>` and the ways directories are combined on Windows are quite new to me. It just goes to show there is something new to learn every day.

10:30am.

```
  outputPrimDocs :: Make ()
  outputPrimDocs = do
    codegenTargets <- asks optionsCodegenTargets
    when (S.member Docs codegenTargets) $ for_ Docs.Prim.primModules $ \docsMod@Docs.Module{..} ->
      writeJSONFile (outputFilename modName "docs.json") docsMod
```

Just studying the code like this is interesting, but I'll want to do something real even if it turns out to be HackerRank exercises.

```
      foreignInclude <- case mn `M.lookup` foreigns of
        Just _
          | not $ requiresForeign m -> do
              return Nothing
          | otherwise -> do
              return $ Just $ Imp.App Nothing (Imp.Var Nothing "require") [Imp.StringLiteral Nothing "./foreign.js"]
```

Oh so it is possible to use the guards like this. This is new to me. Personally, I'd just use an if statement.

...

Ok...

I had enough.

```
  codegen :: CF.Module CF.Ann -> Docs.Module -> ExternsFile -> SupplyT Make ()
  codegen m docs exts = do
    let mn = CF.moduleName m
    lift $ writeJSONFile (outputFilename mn "externs.json") exts
    codegenTargets <- lift $ asks optionsCodegenTargets
    when (S.member CoreFn codegenTargets) $ do
      let coreFnFile = targetFilename mn CoreFn
          json = CFJ.moduleToJSON Paths.version m
      lift $ writeJSONFile coreFnFile json
    when (S.member JS codegenTargets) $ do
      foreignInclude <- case mn `M.lookup` foreigns of
        Just _
          | not $ requiresForeign m -> do
```

It is not realistic for me to just study this code, I want to play around with Haskell now that it has come to this.

10:35am.

```
executables:
  hello-exe:
    main:                Main.hs
    source-dirs:         app
    ghc-options:
    - -threaded
    - -rtsopts
    - -with-rtsopts=-N
    dependencies:
    - hello
```

Oh, `package.yaml` does have a `ghc-options` section.

10:40am.

```
2020-03-18 10:41:01.8779772 [ThreadId 1216] - Source error on cradle initialisation: module `Paths_hello' cannot be found locally

2020-03-18 10:41:01.8779772 [ThreadId 1216] - Cradle did not load succesfully
```

I keep getting this error even though the plugin works. Well, nevermind it.

10:45am. Now it is downloading GHC 8.8.3.

I do not understand why `stack build` failed for me in the hello folder. I had to recreate it.

Oh, goddamit. Stack and Hie are downloading 8.8.3 separately. What the hell?

...I'll cancel the stack download.

Let me find a HackerRank problem.

https://www.hackerrank.com/challenges/botclean?hr_b=1

Let me go with this.

10:50am.

```
module Main where

getList :: Int -> IO [String]
getList n = if n==0 then return [] else do i <- getLine; is <- getList(n-1); return (i:is)

next_move :: String -> [String] -> String
    --logic here

main :: IO()
main = do
    -- Take input
    pos <- getLine
    board <- getList 5
    -- Do I/O stuff here.
    putStrLn $ next_move pos board
```

Oh, it is really nice of them to give me the reader function.

This kind of problem would be absolutely trivial for me in F#, but in Haskell, it will be somewhat challenging.

10:55am. The easiest way to beat this would be to implement A*.

Ugh, Hie is complaining that it wants 8.6.5.

Now an impossible bug happened and it is asking me to report this.

```
panic! (the 'impossible' happened)\n (GHC version 8.6.5 for x86_64-unknown-mingw32):\n\tDynamic linker not initialised\n\nPlease report this as a GHC bug: http://www.haskell.org/ghc/reportabug\n"
```

11am. Goddamit. How do I set the ghc version in the yaml file?

```
The final field is resolver. This tells stack how to build your package: which GHC version to use, versions of package dependencies, and so on. Our value here says to use LTS Haskell version 3.2, which implies GHC 7.10.2 (which is why stack setup installs that version of GHC). There are a number of values you can use for resolver, which we'll cover later.
```

11:05am.

```
resolver: ghc-8.6.5
```

Setting it to this fixes everything.

It is really a good thing that I canceled the download of 8.8.3 and let Hie do it instead.

Now where was I?

11:10am. Right now I am thinking of how to do this problem.

...And am getting annoyed at how more difficult Haskell is making this.

Yeah, this brings back memories. Sigh.

It would be perfectly trivial to do this problem with some help from dictionaries. It would be trivial to implement a dynamic programming solution, but this is one thing Haskell makes annoyingly difficult.

11:15am. Forget memoization or anything else.

What I need here is a logger monad to keep track of the moves taken. Since we are talking only about a 5x5 grid here, I should just be able to do an exhaustive breadth first search. It should not be a problem. Is there a queue in the standard Haskell library somewhere?

11:20am. http://hackage.haskell.org/package/queue-0.1.2/docs/Data-Queue.html

But I can't just import outside packages in HR problems.

...Let me take a break here.

11:35am. Done with the small break. How about I take a big one?

I said yesterday that I would learn enough Haskell to 'do it professionally', but thinking through how I would do the HN problem is showing me just how much pain in the ass programming in Haskell would be.

Last time, back in 2016, I sort of did it all myself, but here if I am going to be doing HN problems seriously, one thing I would like to do is instead look at other people's solutions.

Even if I made up my mind to do this problem, right now I am at a loss on how to do breadth first search in Haskell. Yeah, it is possible, but everything that I can think of seems like a pain.

11:40am. You know what, forget this. Doing these problems is a different sort of challenge from understanding the compiler or the monad transformers.

No doubt HR tier problems would be a lot easier if I was intimately familiar with Haskell's standard library, but for now they are besides the point.

I am going to go back to what I was doing which is studying the compiler.

11:40am. Let me take a real break here. After I am done with breakfast, I'll get back to studying the compiler. Forget everything else. Seriously, what a waste of time Haskell is."