"1:15pm. Done with breakfast and chores. What is next on the list?
1:30pm. Let me start this thing.

First of all, let me go through all the files top to bottom.

That will give me an idea of what I need to do.

...

The editor support for Purescript is actually quite bad. What the hell?

1:35pm.

```
      startPscIdeServer = do
        liftEffect $ logError Info "Starting IDE server"
        rootPath <- liftEffect $ (_.root <<< unwrap) <$> Ref.read state
        settings <- liftEffect $ Ref.read config
        startRes <- startServer' settings rootPath logError logError
        retry logError 6 case startRes of
          { port: Just port, quit } -> do
            loadAll port >>= case _ of
              Left msg -> liftEffect $ logError Info $ "Non-fatal error loading modules: " <> msg
              _ -> pure unit
            liftEffect $ Ref.modify_ (over ServerState $ _ { port = Just port, deactivate = quit }) state
          _ -> pure unit
```

Hover the cursor over the locals and you won't get any type feedback. Actually the editor support here is worse than even for Haskell.

1:40pm.

```
  onCompletion conn $ runHandler "onCompletion" getTextDocUri (getCompletions documents)
  onDefinition conn $ runHandler "onDefinition" getTextDocUri (getDefinition documents)
  onDocumentSymbol conn $ runHandler "onDocumentSymbol" getTextDocUri getDocumentSymbols
  onWorkspaceSymbol conn $ runHandler "onWorkspaceSymbol" (const Nothing) getWorkspaceSymbols

  onReferences conn $ runHandler "onReferences" (const Nothing) (getReferences documents)
  onHover conn $ runHandler "onHover" getTextDocUri (getTooltips documents)
  onCodeAction conn $ runHandler "onCodeAction" getTextDocUri (getActions documents)
  onShutdown conn $ fromAff stopPscIdeServer
```

Ffffhhhhh...

1:45pm. Somehow the Purescript code seems cleaner, but `Main.purs` is killing me. Let me check out the other files. I'll conclude this thing very soon. I am at my limit here.

1:55pm. No, this is not going for me.

Once more let me switch to the next phase.

I can't use this as a template after all. It is too complicated.

I have gained something from this Purescript adventure, but I can't quite grab the thing that I want.

2pm. It is a struggle. I start playing with code and get into it, but now that I am just reading things I am out like a candle once again.

I guess it does not matter what the language is. It is too much to expect to learn just by reading. One has to develop on his own.

In terms of how complete my plan is - on the F# side, the details of the IDE implementation are well known to me. The server itself won't be the problem.

The thing that is still a problem for me is this thing I am writing on - VS Code.

Yes - I know I can just use those handles and move from there, but there is no way I could ever be satisfied with such a meager amount of understanding.

Understanding the PS plugin and compiler implementation might not be necessary, but I definitely need to understand the editor itself. Everything else I can wing.

2:05pm. Right now, let me stop here for a while. I need to do some more chores."